callback functions added and removed unneccessary code
